+++ packages/yourloops/test/integration/use-cases/app-main-layout-visualisation.ts
@@ -32,6 +32,7 @@ import {
   checkPatientHeader,
   type HeaderInfo
 } from '../assert/header.assert'
+import { checkBannerLanguageChangeMobile, checkHcpHeaderMobile, checkCaregiverHeaderMobile, checkPatientHeaderMobile } from '../assert/header-mobile.assert'
 import { checkFooterForCaregiver, checkFooterForHcp, checkFooterForPatient } from '../assert/footer.assert'
 import { checkPatientNavBarAsHcp, checkPatientNavBarAsHcpInPrivateTeam } from '../assert/patient-nav-bar.assert'
 
@@ -50,9 +51,17 @@ export const testAppMainLayoutForHcp = async (appMainLayoutParams: AppMainLayout
   checkFooterForHcp(appMainLayoutParams.footerHasLanguageSelector ?? false)
 }
 
+export const testAppMainLayoutForHcpMobile = async (appMainLayoutParams: AppMainLayoutHcpParams) => {
+  await checkHcpHeaderMobile(appMainLayoutParams.headerInfo)
+}
+
 export const testAppMainLayoutForCaregiver = async (appMainLayoutParams: AppMainLayoutParams) => {
   await checkCaregiverHeader(appMainLayoutParams.loggedInUserFullName)
   checkFooterForCaregiver(appMainLayoutParams.footerHasLanguageSelector ?? false)
+}
+
+export const testAppMainLayoutForCaregiverMobile = async (appMainLayoutParams: AppMainLayoutParams) => {
+  await checkCaregiverHeaderMobile(appMainLayoutParams.loggedInUserFullName)
 }
 
 export const testAppMainLayoutForPatient = async (appMainLayoutParams: AppMainLayoutParams) => {
@@ -60,8 +69,13 @@ export const testAppMainLayoutForPatient = async (appMainLayoutParams: AppMainLa
   checkFooterForPatient(appMainLayoutParams.footerHasLanguageSelector ?? false)
 }
 
+export const testAppMainLayoutForPatientMobile = async (appMainLayoutParams: AppMainLayoutParams) => {
+  await checkPatientHeaderMobile(appMainLayoutParams.loggedInUserFullName)
+}
+
 export const testBannerLanguageUpdate = async () => {
   await checkBannerLanguageChange()
+  await checkBannerLanguageChangeMobile()
 }
 
 export const testPatientNavBarLayoutForHcp = () => {